+++ app/src/views/features/rules/RuleEditor/components/RuleDetailsPanel/constants.ts
@@ -2,27 +2,7 @@ import { ReactNode } from "react";
 import { RuleType } from "types";
 import RULE_TYPES_CONFIG from "config/constants/sub/rule-types";
 import LINKS from "config/constants/sub/links";
-
-export enum ExampleType {
-  USE_TEMPLATE = "use_template",
-  DOWNLOAD_DESKTOP_APP = "download_desktop_app",
-  PLAYGROUND_LINK = "playground_link",
-  UNKNOWN = "unknown",
-}
-
-type UseCaseExample =
-  | {
-      type: ExampleType.USE_TEMPLATE;
-      suggestedTemplateId: string;
-    }
-  | {
-      type: ExampleType.PLAYGROUND_LINK;
-      link: string;
-    }
-  | {
-      type: ExampleType.DOWNLOAD_DESKTOP_APP;
-      link: string;
-    };
+import { ExampleType, UseCaseExample } from "./types";
 
 export const RULE_DETAILS: Record<
   RuleType,
